Add js code to toggle color on hover & click
I used "mouseenter" to handle the color change when the cursor passes
over a square, but I didn't love that it is instantaneous, as it just
leads to scribbles, so I used modifier keys to give the user some
control. To "draw", hold down the shift key. To erase, hold down the
alt/option key. I also added a click listener to give very fine
control, but this actually doesn't work as well, I think because the
mouseenter listener conflicts a little bit. I think you have to enter
the "pixel", and then click for it to work. But, it does turn the
"pixel" on AND off, so that's nice. This allowed me to remove the
:hover css style.