Add functions to change width and clear drawing
My function to clear the drawing area is pretty straight-forward, and
basically does the opposite of the drawing area function (which is now
its own function, too). I toyed with using a page load listener to
initialize the grid, but I'm not sure that's necessary. I might add it
back in, though.

I also don't know if I mentioned in a previous commit my use of
document.stylesheets to get the width of the container and to set
the number of squares per row. This was the only way I could figure to
get the width into the .js file, other than actually setting it there,
but then it would still have to be passed back to the CSS or HTML.
Passing it to the HTML is easy, but points to the same reason why I
didn't do this for the squares, too - I odn't like inline styles in my
HTML. However, this does lead to a minor potential issue in that the
styles are referenced by index, so if the CSS changes, the js file
might need to be updated.