Add a function to validate user-input grid size
Added a function sizeValidation() that is called when the user clicks
the "Set Screen Size" button. I used a recursive function to check the
value entered, and an if statement in the eventhandler, so that if the
user cancels or doesn't enter a number, the grid doesn't clear. If the
value is 0 or negative, the function alerts the user, then calls itself
to prompt the user for a value. If the value is greater than 100, it
alerts that the max is 100, and calls itself. Finally, if the value
entered is not a whole number... This got interesting. The result of
the prompt is a string, and that gets passed back into the rest of the
script as such, but I needed to check that it was a whole number in
string form. Now, isNaN() is a great built-in function that will check
the contents of a string, but it doesn't care about number type (float,
int, etc), so "Four" would be caught, but "4.2" would not. So, I turned
to Number.isInteger(), but obviously that needs a number object,
meaning I had to convert the input.  parseInt() effectively rounds the
number down, returning just the integer portion, so that would always
return true. parseFloat(), however, returns the full float
representation. Additionally, even if the integer is entered in float
notation for some reason (i.e. 4.0), and parseFloat() returns it with
that trailing 0, Number.isInteger(n.0) still evaluates to true. So, I
to convert the input to float, and then check if it's an integer. If
not, alert the user and ask for a whole number.