Add functionality to gradually change opacity
I added two variables, counter and opacity, so that I could track how
often a square has been interacted with. As I have draw and erase
functionality, I also needed to evaluate which was being done. So, a
draw event increments the counter by 1, an erase event decrements it by
one. Opacity is the counter multiplied by 0.1. I also added a check so
that the counter would never actually go above 10 or less than 0, so
that if one were to darken a square more than 10 times, erasing it
would set the opacity to the correct value (i.e. if one interacted with
a square 16 times, the counter would be at 16; if one then tried to
decrease the opacity, the code would set the opacity to 15*0.1 = 1.5
instead of 9*0.1 = 0.9). I left the click functionality as changing the
square to a random color (at the current opacity), because I don't
really care for either the random colors or the opacity change, and
will be removing all of that the next time I work on this project.