Remove all touch controls
Yet another thing that's beyond my abilities right now. I really want
to make this usable on touch screens, but I would need to take a bit of
time to sit down and really figure out the logic of that. I think I was
close, but it might require an entire re-working of the script, and I
just don't have time for that right now. All the code has been pulled,
and I added text to the "instructions" that lets the user know.